refactor(probe): complete arp migration with services and cli hooks
