+++ tools/probe/engine/cli.py
@@ -137,7 +137,6 @@ from .endpoint_addressing import (
     apply_shared_ipv4_rewrite_tail as _apply_shared_ipv4_rewrite_tail,
     default_failure_reasons as _default_failure_reasons,
     _eui64_link_local_ipv6,
-    _lab_arp_alias_ipv4,
 )
 from .protocols import (
     all_stimulus_endpoint_cases as _registry_stimulus_endpoint_cases,
@@ -1631,258 +1630,6 @@ def _probe_plan_with_endpoint_addresses(
                 send["validation"] = send_validation
                 rewritten_udp_sends.append(send)
             updated["udp_sends"] = rewritten_udp_sends
-    elif case_name in {
-        "arp-basic-who-has",
-        "arp-repeat-two-replies",
-        "arp-source-address-preserved",
-        "arp-alias-address-reply",
-        "arp-unicast-request-reply",
-        "arp-padding-reply",
-        "arp-cache-flush-reply",
-        "arp-mac-validation",
-        "arp-spa-variation",
-        "arp-broadcast-filtered-capture",
-    }:
-        # ARP rides Ethernet directly (no IP/UDP), so the lab rewrite touches the
-        # ARP protocol addresses rather than transport IPs: the stimulus resolves
-        # the target endpoint's IPv4 (target protocol address) from its own IPv4
-        # (sender protocol address). The Ethernet source stays the stimulus MAC
-        # and the request is broadcast; the target hardware/protocol addresses the
-        # reply resolves are the target endpoint's on-segment MAC/IPv4. The
-        # capture filter is link-layer (ARP plus the reply opcode), so it carries
-        # no host/IP terms to rewrite. ARP carries no transport source/destination
-        # IPv4 to overwrite, so this branch returns directly rather than falling
-        # into the shared IP-layer validation tail.
-        #
-        # arp-alias-address-reply is the exception: the kernel answers for a
-        # *configured secondary* IPv4 alias, not its primary on-segment address.
-        # The resolved target protocol address (and the expected reply sender
-        # protocol address) is therefore a distinct alias host in the lab subnet,
-        # derived deterministically from the target endpoint's IPv4 and threaded
-        # into the target_service setup/cleanup (alias add/del) and the expected
-        # reply sender-proto.
-        is_alias_case = case_name == "arp-alias-address-reply"
-        is_spa_case = case_name == "arp-spa-variation"
-        is_filtered_capture_case = case_name == "arp-broadcast-filtered-capture"
-        resolved_source_mac = _string_or(source_mac, "")
-        resolved_target_mac = _string_or(target_mac, "")
-        resolved_target_interface = _string_or(target_interface, "")
-        resolved_ipv4 = (
-            _lab_arp_alias_ipv4(target_ipv4, source_ipv4)
-            if is_alias_case
-            else target_ipv4
-        )
-        # arp-spa-variation is the sender-side exception: the who-has carries an
-        # ALTERNATE sender protocol address (SPA), distinct from the stimulus
-        # endpoint's primary on-segment IPv4 (``source_ipv4``). The SPA is a single
-        # source of truth: the request's sender_protocol_addr and the expected
-        # reply target_protocol_addr are both the alternate SPA. Derive it as a
-        # distinct host on the lab segment (reuse the alias-style host derivation,
-        # which already avoids the source/target/router/broadcast hosts).
-        sender_proto = (
-            _lab_arp_alias_ipv4(source_ipv4, target_ipv4)
-            if is_spa_case
-            else source_ipv4
-        )
-        updated["source_ipv4"] = source_ipv4
-        updated["destination_ipv4"] = resolved_ipv4
-        updated["expected_reply_source_ipv4"] = resolved_ipv4
-        updated["expected_reply_destination_ipv4"] = sender_proto
-        updated["sender_protocol_addr"] = sender_proto
-        updated["target_protocol_addr"] = resolved_ipv4
-        if resolved_source_mac:
-            updated["sender_hardware_addr"] = resolved_source_mac
-            updated["ethernet_source"] = resolved_source_mac
-        if resolved_target_mac and case_name == "arp-unicast-request-reply":
-            updated["ethernet_destination"] = resolved_target_mac
-        if is_alias_case:
-            updated["alias_ipv4"] = resolved_ipv4
-        if is_spa_case:
-            updated["alt_sender_ipv4"] = sender_proto
-        target_service = dict(
-            json_object(updated.get("target_service", {}), "probe_plan.target_service")
-        )
-        target_service.update(
-            {
-                # The kernel's primary on-segment address is the endpoint IPv4; the
-                # alias is added on top of it, so the interface still binds the
-                # endpoint address.
-                "bind_ipv4": target_ipv4,
-                "source_ipv4": source_ipv4,
-                "target_protocol_addr": resolved_ipv4,
-            }
-        )
-        if resolved_target_mac:
-            target_service["target_hardware_addr"] = resolved_target_mac
-            repeat = target_service.get("repeat")
-            if isinstance(repeat, dict) and isinstance(repeat.get("sends"), list):
-                repeat["sends"] = [
-                    {
-                        **json_object(send, "probe_plan.target_service.repeat.send"),
-                        "sender_hardware_addr": resolved_target_mac,
-                    }
-                    for send in repeat["sends"]
-                    if isinstance(send, Mapping)
-                ]
-                target_service["repeat"] = repeat
-        if resolved_target_interface:
-            target_service["interface"] = resolved_target_interface
-        if is_alias_case:
-            # Target setup adds the secondary alias to the private interface (and
-            # removes it during cleanup); rewrite the alias onto the lab segment so
-            # the live setup configures the on-segment alias the who-has resolves.
-            target_service["alias_ipv4"] = resolved_ipv4
-            target_service["alias_address"] = True
-        if is_spa_case:
-            # For live execution the target kernel may need the alternate SPA
-            # configured as a secondary sender address so it accepts/answers the
-            # who-has; thread the on-segment SPA into the target service so setup
-            # adds (and cleanup removes) the secondary sender address.
-            target_service["alt_sender_ipv4"] = sender_proto
-            target_service["alt_sender_address"] = True
-        if is_filtered_capture_case:
-            decoy_sender_ipv4 = _lab_arp_alias_ipv4(target_ipv4, source_ipv4)
-            decoy_target_ipv4 = _lab_arp_alias_ipv4(source_ipv4, target_ipv4)
-
-            def _rewrite_decoy_arp_event(raw_event: object) -> JSONObject:
-                event = dict(json_object(raw_event, "probe_plan.decoy_arp_event"))
-                event["sender_protocol_addr"] = decoy_sender_ipv4
-                event["target_protocol_addr"] = decoy_target_ipv4
-                if resolved_target_mac:
-                    event["ethernet_source"] = resolved_target_mac
-                    event["sender_hardware_addr"] = resolved_target_mac
-                if resolved_source_mac:
-                    event["ethernet_destination"] = resolved_source_mac
-                    event["target_hardware_addr"] = resolved_source_mac
-                return event
-
-            if isinstance(updated.get("decoy_arp_event"), dict):
-                updated["decoy_arp_event"] = _rewrite_decoy_arp_event(
-                    updated["decoy_arp_event"]
-                )
-            target_service["decoy_arp_event"] = updated.get("decoy_arp_event")
-        updated["target_service"] = target_service
-        arp_validation = dict(
-            json_object(updated.get("validation", {}), "probe_plan.validation")
-        )
-        arp_validation["sender_protocol_addr"] = resolved_ipv4
-        # The reply is addressed back to the request's sender protocol address: the
-        # alternate SPA for arp-spa-variation, the stimulus IPv4 otherwise.
-        arp_validation["target_protocol_addr"] = sender_proto
-        if resolved_target_mac:
-            arp_validation["sender_hardware_addr"] = resolved_target_mac
-            arp_validation["ethernet_source"] = resolved_target_mac
-            if "provider_mac" in arp_validation:
-                arp_validation["provider_mac"] = resolved_target_mac
-        if resolved_source_mac:
-            arp_validation["target_hardware_addr"] = resolved_source_mac
-            arp_validation["ethernet_destination"] = resolved_source_mac
-        updated["validation"] = arp_validation
-        if case_name == "arp-unicast-request-reply":
-            # The unicast case sends the request directly to the known target MAC
-            # rather than the broadcast address. Pin the request's Ethernet
-            # destination to the target endpoint's hardware address (the resolved
-            # target MAC the reply also carries) so the request stays unicast and
-            # the target MAC remains a single source of truth across the request
-            # frame, the target service, and the is-at validation contract.
-            target_mac = target_service.get("target_hardware_addr")
-            if isinstance(target_mac, str) and target_mac:
-                updated["ethernet_destination"] = target_mac
-        if case_name == "arp-cache-flush-reply":
-            # The cache-flush case carries an explicit pre-stimulus neighbor flush
-            # (a `flush_neighbor` marker plus the setup/cleanup `ip neigh` commands)
-            # both at the top level (read by the endpoint via the flattened plan)
-            # and mirrored into the target service. The documentation-space commands
-            # name the stimulus IPv4; rewrite that onto the lab segment so the live
-            # setup flushes the querier the on-segment who-has actually resolves.
-            doc_stimulus_ipv4 = str(plan.get("source_ipv4", ""))
-
-            def _rewrite_flush(commands: object) -> object:
-                if not isinstance(commands, list):
-                    return commands
-                rewritten = [
-                    command.replace(doc_stimulus_ipv4, source_ipv4)
-                    if isinstance(command, str) and doc_stimulus_ipv4
-                    else command
-                    for command in commands
-                ]
-                if resolved_target_interface:
-                    old_interface = _string_or(
-                        updated.get("neighbor_flush_interface"),
-                        _string_or(target_service.get("neighbor_flush_interface"), "eth0"),
-                    )
-                    rewritten = [
-                        command.replace(
-                            f" dev {old_interface}",
-                            f" dev {resolved_target_interface}",
-                        )
-                        if isinstance(command, str)
-                        else command
-                        for command in rewritten
-                    ]
-                return rewritten
-
-            for key in ("neighbor_flush_commands", "neighbor_flush_cleanup_commands"):
-                if key in updated:
-                    updated[key] = _rewrite_flush(updated[key])
-                if key in target_service:
-                    target_service[key] = _rewrite_flush(target_service[key])
-            if resolved_target_interface:
-                updated["neighbor_flush_interface"] = resolved_target_interface
-                target_service["neighbor_flush_interface"] = resolved_target_interface
-            updated["target_service"] = target_service
-        # arp-repeat-two-replies carries a per-send array: rewrite each who-has
-        # send's ARP protocol addresses and is-at validation contract onto the lab
-        # segment so every send resolves the same target and each decoded is-at
-        # reply is validated against its own send. The two sends share the target
-        # (the case point is a repeated who-has for one target), so they all
-        # resolve the same on-segment protocol addresses.
-        arp_sends = updated.get("arp_sends")
-        if isinstance(arp_sends, list):
-            rewritten_arp_sends: list[JSONObject] = []
-            for raw_send in arp_sends:
-                send = dict(json_object(raw_send, "probe_plan.arp_send"))
-                send["source_ipv4"] = source_ipv4
-                send["destination_ipv4"] = target_ipv4
-                send["expected_reply_source_ipv4"] = target_ipv4
-                send["expected_reply_destination_ipv4"] = source_ipv4
-                send["sender_protocol_addr"] = source_ipv4
-                send["target_protocol_addr"] = target_ipv4
-                if resolved_source_mac:
-                    send["sender_hardware_addr"] = resolved_source_mac
-                    send["ethernet_source"] = resolved_source_mac
-                send_validation = dict(
-                    json_object(send.get("validation", {}), "probe_plan.arp_send.validation")
-                )
-                send_validation["sender_protocol_addr"] = target_ipv4
-                send_validation["target_protocol_addr"] = source_ipv4
-                if resolved_target_mac:
-                    send_validation["sender_hardware_addr"] = resolved_target_mac
-                    send_validation["ethernet_source"] = resolved_target_mac
-                if resolved_source_mac:
-                    send_validation["target_hardware_addr"] = resolved_source_mac
-                    send_validation["ethernet_destination"] = resolved_source_mac
-                send["validation"] = send_validation
-                rewritten_arp_sends.append(send)
-            updated["arp_sends"] = rewritten_arp_sends
-        live_rewrite: JSONObject = {
-            "source": rewrite_source,
-            "stimulus_ipv4": source_ipv4,
-            "target_ipv4": target_ipv4,
-        }
-        if is_alias_case:
-            live_rewrite["alias_ipv4"] = resolved_ipv4
-        if is_filtered_capture_case:
-            decoy_arp_event = updated.get("decoy_arp_event")
-            if isinstance(decoy_arp_event, dict):
-                live_rewrite["decoy_sender_ipv4"] = decoy_arp_event.get(
-                    "sender_protocol_addr"
-                )
-                live_rewrite["decoy_target_ipv4"] = decoy_arp_event.get(
-                    "target_protocol_addr"
-                )
-        updated["live_address_rewrite"] = live_rewrite
-        return updated
     return _apply_shared_ipv4_rewrite_tail(
         updated,
         case_name=case_name,
@@ -2060,41 +1807,6 @@ def _failure_reasons_for_case(case_name: str) -> list[str]:
             FAILURE_TARGET_SETUP_FAILED,
         ]
     if case_name == "ttl-expired":
-        return [
-            FAILURE_TIMEOUT,
-            FAILURE_WRONG_PEER,
-            FAILURE_WRONG_PAYLOAD,
-            FAILURE_DECODE_FAILED,
-        ]
-    if case_name in {
-        "arp-alias-address-reply",
-        "arp-cache-flush-reply",
-        "arp-spa-variation",
-        "arp-broadcast-filtered-capture",
-    }:
-        # The alias case adds (and removes) a secondary IPv4 on the target
-        # interface as part of target setup; the cache-flush case flushes the
-        # neighbor cache before the stimulus as part of target setup; the
-        # spa-variation case may configure a secondary *sender* IPv4 so the kernel
-        # accepts the reply addressed to the alternate SPA; the filtered-capture
-        # case may emit a decoy ARP setup event. Any setup step failing is a
-        # distinct failure mode on top of the shared ARP reasons.
-        return [
-            FAILURE_TIMEOUT,
-            FAILURE_WRONG_PEER,
-            FAILURE_WRONG_PAYLOAD,
-            FAILURE_DECODE_FAILED,
-            FAILURE_TARGET_SETUP_FAILED,
-        ]
-    if case_name in {
-        "arp-resolution",
-        "arp-basic-who-has",
-        "arp-repeat-two-replies",
-        "arp-source-address-preserved",
-        "arp-unicast-request-reply",
-        "arp-padding-reply",
-        "arp-mac-validation",
-    }:
         return [
             FAILURE_TIMEOUT,
             FAILURE_WRONG_PEER,
